requirements.txt
aiohttp~=3.10.10
pandas~=2.2.3
python-dotenv~=1.0.1
openpyxl~=3.1.5
asyncpg